File Handling
File Handling in C++
File handling in C++ is a mechanism to create, read, write, and manipulate files on a disk. It allows data to be stored permanently in secondary memory, such as a hard disk. The process involves several steps: naming a file, opening a file, writing data to the file, reading data from the file, and closing the file